god-class.tsまわりのバグを修正
